create a directory for ore stuff
There is an existing executable made my the Makefile in applications
called `ore`. To avoid clashes with this executable, the new directory
is named `ore_dir` instead of the more desirable name `ore`. Once the
work in `ore.c` is merged into `ore_dir`, we can remove `ore.c` and
rename the directory.