never do interesting work inside an assert
The implementation of assert means it is not guaranteed that arguments
to assert are evaluated -- in particular, when debug mode is turned off,
arguments are never evaluated. That means that putting allocations or
initializations inside a call to assert is very dangerous. This commit
replaces all instances of that pattern with one that stores results in a
temporary variable and asserts the value of the temporary variable,
instead.